Do some moving about and tidying.
- Methods in TownySettings moved to DatabaseConfig class.
- Remove unused playermap and related method in TownySettings.
